src/test/java/com/crypto/trading/signal/service/OneDaySignalServiceTest.java
package com.crypto.trading.signal.service;

import com.crypto.trading.signal.entity.OneDay;
import com.crypto.trading.signal.model.TradingSignal;
import com.crypto.trading.signal.repository.OneDayRepository;
import org.junit.jupiter.api.Test;
import org.junit.jupiter.api.extension.ExtendWith;
import org.mockito.InjectMocks;
import org.mockito.Mock;
import org.mockito.junit.jupiter.MockitoExtension;

import java.util.List;

import static org.assertj.core.api.Assertions.assertThat;
import static org.mockito.BDDMockito.given;

@ExtendWith(MockitoExtension.class)
public class OneDaySignalServiceTest {

    @InjectMocks
    OneDaySignalService service;
    @Mock
    private OneDayRepository repository;

    @Test
    void willFindAllRecords() {

        given(repository.findAll()).willReturn(List.of(new OneDay("BTCUSDT", TradingSignal.BUY, 0)));

        var result = service.getAll();

        assertThat(result).isNotEmpty();

    }
}
